==> examples/01_bert_train.ipynb <==
import os
os.chdir('/content')

!git clone https://github.com/rightlit/nlp2

%tensorflow_version 1.x

!pip list | egrep "konlpy|torch|tensorflow"

!pip install konlpy

import os
os.chdir('/content/nlp2/examples')

import urllib.request

# NSMC dataset download
print('NSMC dataset downloading...')
urllib.request.urlretrieve("https://raw.githubusercontent.com/e9t/nsmc/master/ratings_train.txt", filename="ratings_train.txt")
urllib.request.urlretrieve("https://raw.githubusercontent.com/e9t/nsmc/master/ratings_test.txt", filename="ratings_test.txt")   
#print('download completed')

# NSMC 데이터
import pandas as pd

df_train = pd.read_csv('./ratings_train.txt', names=['id', 'document','label'], header=0, delimiter = '\t', quoting = 3, encoding='UTF-8')
df_test = pd.read_csv('./ratings_test.txt', names=['id', 'document','label'], header=0, delimiter = '\t', quoting = 3, encoding='UTF-8')

df_train.head()

with open('./train.txt', 'w') as f1:
  for text, label in zip(df_train['document'], df_train['label']):
    #f1.writelines(text + "\u241E" + str(label) + "\n")
    #f1.writelines(text + "\u241E" + str(label) + "\n")
    f1.writelines("%s\u241E%d\n" %(text, label))

with open('./test.txt', 'w') as f2:
  for text, label in zip(df_test['document'], df_test['label']):
    #f2.writelines(text + "\u241E" + str(label) + "\n")
    f2.writelines("%s\u241E%d\n" %(text, label))


!head ./train.txt

!head ./test.txt

!wc -l ./*.txt

!pwd

!mkdir -p ./tune-ckpt

!wget https://storage.googleapis.com/bert_models/2018_11_23/multi_cased_L-12_H-768_A-12.zip

!unzip multi_cased_L-12_H-768_A-12.zip

!ls -l ./multi_cased_L-12_H-768_A-12/*

!ls -l 

# fine tuning (num_labels = 2)
!python ./tune_utils.py --model_name bert \
  --train_corpus_fname ./train.txt \
  --test_corpus_fname ./test.txt \
  --vocab_fname ./multi_cased_L-12_H-768_A-12/vocab.txt \
  --pretrain_model_fname ./multi_cased_L-12_H-768_A-12/bert_model.ckpt \
  --config_fname ./multi_cased_L-12_H-768_A-12/bert_config.json \
  --model_save_path ./tune-ckpt \
  --num_labels 2

# train_corpus_fname=train.txt
# test_corpus_fname=test.txt
!python ./01_bert_train.py --train_corpus_fname=train.txt.1 --test_corpus_fname=test.txt.1

!cp /content/drive/My\ Drive/work/nlp/tune-ckpt/* ./tune-ckpt/

!ls -l ./tune-ckpt

!python ./01_bert_test.py
