==> examples/02_chatbot_kogpt2_train.ipynb <==
import os
os.chdir('/content')

!git clone https://github.com/rightlit/nlp2

%tensorflow_version 1.x

import tensorflow as tf
tf.__version__

!python -V

!pip list | egrep "kobert|transformers|tokenizers"

!pip install transformers
!pip install kogpt2-transformers

import os
os.chdir('/content/nlp2/examples/')

#!cp /content/drive/My\ Drive/work/nlp/ChatbotData.csv ./

!head ./ChatbotData.csv

import pandas as pd
import re

df = pd.read_csv('./ChatbotData.csv')
A_data = df['A'].tolist()
QA_data = list()
for i, q_text in enumerate(df['Q']):
    qa_text = q_text + '    ' + A_data[i]
    #qa_text = qa_text.replace("[^ㄱ-ㅎㅏ-ㅣ가-힣 ]","")
    #qa_text = qa_text.replace("[ㄱ-ㅎㅏ-ㅣ가-힣 ]","")
    qa_text = re.sub("[ㅠㅜ]","", qa_text)
    QA_data.append(qa_text)

df['QA'] = QA_data
df['QA'].to_csv('./chatbot_dialog.txt', index=False, header=False)

#with open("chatbot_dialog.txt", "w") as f:
#    for line in df['QA']:
#        f.writeline(line)

!ls -l

!wc -l ./chatbot_dialog.txt

!head -n 20 ./chatbot_dialog.txt

os.environ['CUDA_LAUNCH_BLOCKING'] = "1"

#torch.backends.cudnn.enabled = False 

!python ./02_chatbot_kogpt2_train.py

#!cp /content/drive/MyDrive/work/nlp/kogpt2-chatbot-dialog.pth ./

!python ./02_chatbot_kogpt2_test.py
